fix: use unit title to get unit download file id
